+++ docs/CHANGELOG.rst
@@ -9,6 +9,10 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 Unreleased
 ==========
 
+Added
+-----
+- Data, sample and collection duplication endpoints
+
 ===================
 10.0.3 - 2019-03-07
 ===================